refactor: cleanup configurator index.ts
