+++ util/config/configurator/package.json
@@ -3,7 +3,8 @@
   "module": "index.ts",
   "type": "module",
   "scripts": {
-    "start": "bun index.ts"
+    "start": "bun index.ts",
+    "test": "bun test tests/index.test.ts"
   },
   "devDependencies": {
     "@types/bun": "latest"
